fix error 1. missing meun header from calendar 2. inventory can add 0.5 number 3. add shoppinlist if not present in inventory then add to inventory

+++ server/controllers/mealPlan.controller.js
@@ -1,7 +1,12 @@
-import e from "express";
 import MealPlan from "../models/mealPlan.model.js";
 import Recipe from "../models/recipe.model.js";
 import { successResponse, errorResponse } from "../utils/apiResponse.js";
+import PantryItem from "../models/pantryItem.model.js";
+import RecipeIngredient from "../models/recipeIngredient.model.js";
+import Ingredient from "../models/Ingredient.model.js";
+import RedisHelper from "../helpers/redisHelper.js";
+import shoppingList from "../models/shoppingList.model.js";
+
 // Get all meal plans
 export const getAllMealPlans = async (req, res) => {
     try {
@@ -51,8 +56,73 @@ export const createMealPlan = async (req, res) => {
             meal_type: request.meal_type,
             serving_date: request.serving_date,
         })
-        res.json(successResponse(newMealPlan));
+
+        const recipe = await Recipe.findById(newMealPlan.recipe_id);
+        // check pantry inventory for recipe ingredients is enough
+        const pantryCacheItems = await RedisHelper.getCache(RedisHelper.REDIS_KEYS.PANTRY_ITEMS(req.auth.user_id));
+        const recipeIngredients = await RecipeIngredient.find({ recipe_id: recipe._id });
+        const notEnoughItems = [];
+        for (const needIngredients of recipeIngredients) {
+            const recordNotEnoughItem = {};
+
+            // find from pantryItems cache first
+            let pantryItem = null;
+            let ingredient = null;
+            if (pantryCacheItems) {
+                pantryItem = pantryCacheItems.find(item => item.ingredient_id.toString() === needIngredients.ingredient_id.toString());
+            }
+            ingredient = await Ingredient.findById(needIngredients.ingredient_id);
+
+
+            const totalNeededQuantity = pantryItem ? pantryItem.item_planned + needIngredients.quantity : needIngredients.quantity;
+            if (!pantryItem || pantryItem.quantity < totalNeededQuantity) {
+                recordNotEnoughItem.ingredient_id = needIngredients.ingredient_id;
+                recordNotEnoughItem.name = ingredient ? ingredient.name : pantryItem ? pantryItem.name : "Unknown";
+                recordNotEnoughItem.required_quantity = totalNeededQuantity - (pantryItem ? pantryItem.quantity : 0);
+                recordNotEnoughItem.ingredient_name = pantryItem ? pantryItem.name : ingredient.name;
+                recordNotEnoughItem.available_quantity = pantryItem ? pantryItem.quantity : 0;
+                notEnoughItems.push(recordNotEnoughItem);
+
+                // add to shopping list
+                const existingShoppingListItem = await shoppingList.findOne({
+                    user_id: req.auth.user_id,
+                    checked: false,
+                    ingredient_id: needIngredients.ingredient_id
+                });
+                if (existingShoppingListItem) {
+                    existingShoppingListItem.quantity += needIngredients.quantity;
+                    await existingShoppingListItem.save();
+                } else {
+                    await shoppingList.create({
+                        user_id: req.auth.user_id,
+                        ingredient_id: needIngredients.ingredient_id,
+                        quantity: recordNotEnoughItem.required_quantity,
+                        checked: false,
+                        unit: ingredient ? ingredient.default_unit : ""
+                    });
+                }
+            }
+
+            // set planned quantity in pantry cache
+            if (pantryItem) {
+                pantryItem.item_to_buy += needIngredients.quantity;
+                pantryItem.item_planned += needIngredients.quantity;
+                RedisHelper.setCache(RedisHelper.REDIS_KEYS.PANTRY_ITEMS(req.auth.user_id), pantryCacheItems, 604800);
+            }
+        }
+
+        const resultMealPlan = {
+            id: newMealPlan._id,
+            recipe_id: recipe.recipe_id,
+            image_url: recipe.image || null,
+            meal_name: recipe.meal_name,
+            meal_type: newMealPlan.meal_type,
+            serving_date: newMealPlan.serving_date,
+            notEnoughItems: notEnoughItems
+        };
+        res.json(successResponse(resultMealPlan));
     } catch (err) {
+        console.error(err);
         res.json(errorResponse({ message: err.message }));
     }
 };
@@ -75,8 +145,29 @@ export const updateMealPlan = async (req, res) => {
 // Delete a meal plan
 export const deleteMealPlan = async (req, res) => {
     try {
-        const deletedMealPlan = await MealPlan.findByIdAndDelete(req.params.id);
-        if (!deletedMealPlan) return res.json(errorResponse({ message: 'Meal plan not found' }));
+        // update pantry planned items before deleting meal plan
+        const mealPlan = await MealPlan.findById(req.params.id);
+        if (mealPlan) {
+            const recipeIngredients = await RecipeIngredient.find({ recipe_id: mealPlan.recipe_id });
+            const pantryCacheItems = await RedisHelper.getCache(RedisHelper.REDIS_KEYS.PANTRY_ITEMS(req.auth.user_id));
+            for (const needIngredients of recipeIngredients) {
+                // find from pantryItems cache first
+                let pantryItem = null;
+                if (pantryCacheItems) {
+                    pantryItem = pantryCacheItems.find(item => item.ingredient_id.toString() === needIngredients.ingredient_id.toString());
+                }
+                // reduce planned quantity in pantry cache
+                if (pantryItem) {
+                    pantryItem.item_to_buy -= needIngredients.quantity;
+                    pantryItem.item_planned -= needIngredients.quantity;
+                    if (pantryItem.item_planned < 0) pantryItem.item_planned = 0;
+                    RedisHelper.setCache(RedisHelper.REDIS_KEYS.PANTRY_ITEMS(req.auth.user_id), pantryCacheItems, 604800);
+                }
+            }
+            await MealPlan.findByIdAndDelete(req.params.id);
+        } else {
+            return res.json(errorResponse({ message: 'Meal plan not found' }));
+        }
         res.json(successResponse({ message: 'Meal plan deleted' }));
     } catch (err) {
         res.json(errorResponse({ message: err.message }));
